Move environment work directory from `~/$SANDBOX/` to `~/w/` (breaking)
Having the work directory at a fixed known location in the environment is
a bit more convenient in many cases.

There's also an obscure reason why it helps Docker volume mounts. When you do a
Docker volume mount on a non-existing directory, it's owned by root. When you
do a Docker volume mount on a directory that exists in the image, it takes on
that directory's ownership. See <https://github.com/moby/moby/issues/2259>.
It's hard to make the directory exist in the image when you can't predict its
name, so this change will help.