Separate environment initialization from execution
This is a fairly large refactor.

Before, `Runner::run()` could be used to unpack seeds and initialize even an
existing environment. This didn't necessarily make sense. For example, it would
install packages without uninstalling what was there before. It's also
potentially difficult to support, like if the new packages require a different
Docker base image, you're out of luck.

Now, the initialization can only happen during `Runner::create()` or
`Runner::reset()`.

As a result of this change, it may be more expensive to rebuild packages. We'll
need new ways to address that later.